+++ namespace_cache.go
